major changes

+++ books/books-router.js
@@ -1,9 +1,13 @@
-const booksRouter = require('express').Router();
-const Book = require('../models/book');
+const booksRouter = require('express').Router()
+const Book = require('../models/book')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 booksRouter.get('/', async (req, res, next) => {
   try {
-    const books = await Book.find({})
+    const books = await Book.find({}).populate('user', {
+      username: 1
+    })
     if (books) {
       res.json(books.map((allBooks) => allBooks.toJSON()))
     } else {
@@ -12,7 +16,7 @@ booksRouter.get('/', async (req, res, next) => {
   } catch (exception) {
     next(exception)
   }
-});
+})
 
 booksRouter.get('/:id', async (req, res, next) => {
   try {
@@ -26,11 +30,16 @@ booksRouter.get('/:id', async (req, res, next) => {
   } catch (exception) {
       next(exception)
   }
-});
+})
 
 booksRouter.post('/', async (req, res, next) => {
   const body = req.body;
-  // const books = await Book.find({})
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+
+  if (!req.token || !decodedToken.id || req.token === null) {
+    return res.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
   const book = new Book({
     title: body.title,
@@ -38,7 +47,8 @@ booksRouter.post('/', async (req, res, next) => {
     genre: body.genre,
     published: body.published,
     dateread: body.dateread,
-  });
+    user: user._id
+  })
 
   try {
     if (
@@ -49,7 +59,8 @@ booksRouter.post('/', async (req, res, next) => {
       body.dateread !== undefined
       ) {
         const savedBook = await book.save();
-        // books = books.concat(savedBook._id)
+        user.books = user.books.concat(savedBook._id)
+        await user.save()
         res.json(savedBook.toJSON)
       } else {
         res.status(400).send('Bad request. Information missing.')
@@ -57,10 +68,23 @@ booksRouter.post('/', async (req, res, next) => {
   } catch (exception) {
     next(exception)
   }
-});
+})
 
 booksRouter.delete('/:id', async (req, res, next) => {
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
   const book = await Book.findById(req.params.id);
+
+  if (user._id.toString() !== book.user.toString()) {
+    return res.status(400).json({ error: 'invalid user' })
+  }
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' })
+  }
 
   try {
     await Book.findByIdAndRemove(req.params.id)
@@ -68,7 +92,7 @@ booksRouter.delete('/:id', async (req, res, next) => {
   } catch (exception) {
       next(exception)
   }
-});
+})
 
 booksRouter.put('/:id', async (req, res, next) => {
   const body = req.body;
@@ -89,6 +113,6 @@ booksRouter.put('/:id', async (req, res, next) => {
       console.log('exception', exception)
       next(exception)
   }
-});
+})
 
-module.exports = booksRouter;
+module.exports = booksRouter
